# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  firstName: string;
  lastName: string;
  username: string;
  email: string;
  bio?: string;
  createdAt: string;
}

export interface RegisterRequest {
  firstName: string;
  lastName: string;
  username: string;
  email: string;
  password: string;
}

export interface LoginRequest {
  /** Email or username */
  identifier: string;
  password: string;
}

export interface UpdateProfileRequest {
  firstName?: string;
  lastName?: string;
  bio?: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UserResponse {
  user: User;
}

export interface Devotional {
  id: number;
  title: string;
  scriptureRef: string;
  scriptureText: string;
  body: string;
  publishDate: string;
  createdAt: string;
}

export interface CreateDevotionalRequest {
  title: string;
  scriptureRef: string;
  scriptureText: string;
  body: string;
}

export interface DevotionalsResponse {
  devotionals: Devotional[];
}

export interface DevotionalResponse {
  devotional: Devotional;
}

export interface ReadingPlanDay {
  id: number;
  planId: number;
  dayNumber: number;
  passage: string;
  reflection?: string;
}

export interface ReadingPlan {
  id: number;
  title: string;
  description?: string;
  createdAt: string;
  days: ReadingPlanDay[];
}

export interface ReadingPlansResponse {
  plans: ReadingPlan[];
}

export interface ReadingProgressItem {
  id: number;
  userId: number;
  planDayId: number;
  completed: boolean;
  completedAt: string;
}

export interface ReadingProgressResponse {
  progress: ReadingProgressItem[];
}

export interface ReadingProgressItemResponse {
  progress: ReadingProgressItem;
}

export interface PrayerRequestItem {
  id: number;
  content: string;
  isAnonymous: boolean;
  prayCount: number;
  createdAt: string;
  author?: string | null;
  isOwn: boolean;
  hasPrayed: boolean;
}

export interface PrayersResponse {
  prayers: PrayerRequestItem[];
}

export interface CreatePrayerRequest {
  content: string;
  isAnonymous?: boolean;
}

export interface PrayToggleResult {
  hasPrayed: boolean;
}

export interface EventItem {
  id: number;
  title: string;
  description?: string;
  location?: string;
  startsAt: string;
  createdAt: string;
  isRsvped: boolean;
}

export interface EventsResponse {
  events: EventItem[];
}

export interface CreateEventRequest {
  title: string;
  description?: string;
  location?: string;
  startsAt: string;
}

export interface RsvpToggleResult {
  isRsvped: boolean;
}

export interface AnswerQuestionRequest {
  answer: string;
}

export type ChatRoom = typeof ChatRoom[keyof typeof ChatRoom];


export const ChatRoom = {
  friends: 'friends',
  anonymous: 'anonymous',
} as const;

export interface ChatMessageItem {
  id: number;
  room: string;
  userId: number;
  displayName: string;
  content: string;
  createdAt: string;
}

export interface ChatMessagesResponse {
  messages: ChatMessageItem[];
}

export interface SendChatMessageRequest {
  content: string;
  displayName: string;
}

export interface ChatMessageResponse {
  message: ChatMessageItem;
}

export interface QuestionItem {
  id: number;
  userId: number;
  ticketCode: string;
  content: string;
  answer?: string | null;
  createdAt: string;
  answeredAt?: string | null;
}

export interface QuestionsResponse {
  questions: QuestionItem[];
}

export interface SubmitQuestionRequest {
  content: string;
}

export interface QuestionResponse {
  question: QuestionItem;
}