# lib/api-zod/src/generated/types/createDevotionalRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateDevotionalRequest {
  title: string;
  scriptureRef: string;
  scriptureText: string;
  body: string;
}